+++ metadata/metadata_service/api/healthcheck.py
@@ -6,7 +6,7 @@ from typing import Dict, Tuple
 from flasgger import swag_from
 from flask_restful import Resource
 
-from metadata_service.proxy import get_proxy_client
+from metadata_service.config import get_proxy_client
 
 
 class HealthcheckAPI(Resource):
