+++ metadata/metadata_service/api/system.py
@@ -7,7 +7,7 @@ from typing import Iterable, Mapping, Union
 from flasgger import swag_from
 from flask_restful import Resource
 
-from metadata_service.proxy import get_proxy_client
+from metadata_service.config import get_proxy_client
 
 
 class Neo4jDetailAPI(Resource):
