+++ metadata/metadata_service/api/tag.py
@@ -10,7 +10,7 @@ from flask import current_app as app
 from flask_restful import Resource, fields, marshal
 
 from metadata_service.exception import NotFoundException
-from metadata_service.proxy import get_proxy_client
+from metadata_service.config import get_proxy_client
 from metadata_service.proxy.base_proxy import BaseProxy
 
 tag_fields = {
